pyproject.toml
[project]
name = "insightql"
version = "0.1.0"
description = "Agente Analítico LLM para Catálogo de Moda - Supabase con MCP + LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LangGraph & LangChain
    "langgraph>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    # Supabase
    "supabase>=2.0.0",
    # Utilities
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    # Observability - LangSmith
    "langsmith>=0.1.0",
    "jsonpointer>=3.0.0",
    "streamlit>=1.53.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
